move option parsing into its own mod
