+++ src/main.rs
@@ -16,6 +16,7 @@ mod fastly_uap;
 mod geo;
 mod handler;
 mod memory;
+mod opts;
 
 use anyhow::anyhow;
 use backend::Backends;
@@ -37,16 +38,16 @@ use hyper::{
     Server,
 };
 use notify::{watcher, DebouncedEvent, RecursiveMode, Watcher};
+use opts::Opts;
 use rustls::internal::pemfile;
 use std::{
     collections::HashMap,
-    error::{Error, Error as StdError},
+    error::Error,
     fs::{self, File},
     io::BufReader,
     path::{Path, PathBuf},
     pin::Pin,
     process::exit,
-    str::FromStr,
     sync::{mpsc::channel, Arc, RwLock},
     time::{Duration, SystemTime},
 };
@@ -61,56 +62,6 @@ use wasmtime::{Engine, Module, Store};
 pub type BoxError = Box<dyn Error + Send + Sync + 'static>;
 type Backend = Vec<(String, String)>;
 type Dictionary = Vec<(String, HashMap<String, String>)>;
-
-fn parse_key_value<T, U>(s: &str) -> Result<(T, U), Box<dyn StdError>>
-where
-    T: FromStr,
-    T::Err: StdError + 'static,
-    U: FromStr,
-    U::Err: StdError + 'static,
-{
-    let pos = s
-        .find(':')
-        .ok_or_else(|| format!("invalid KEY:value: no `:` found in `{}`", s))?;
-    Ok((s[..pos].parse()?, s[pos + 1..].parse()?))
-}
-
-fn parse_dictionary(s: &str) -> Result<(String, HashMap<String, String>), Box<dyn StdError>> {
-    let (name, v) = parse_key_value::<String, String>(s)?;
-    let dict: Result<HashMap<String, String>, Box<dyn StdError>> =
-        v.split(',').try_fold(HashMap::default(), |mut res, el| {
-            let pos = el
-                .find('=')
-                .ok_or_else(|| format!("invalid KEY=value: no `=` found in `{}`", el))?;
-            res.insert(el[..pos].parse()?, el[pos + 1..].parse()?);
-            Ok(res)
-        });
-    Ok((name, dict?))
-}
-
-/// ⏱️  A local Fastly Compute@Edge runtime emulator
-#[derive(Debug, StructOpt)]
-struct Opts {
-    /// Path to a Fastly Compute@Edge .wasm file
-    #[structopt(long, short, default_value = "bin/main.wasm")]
-    wasm: PathBuf,
-    /// Port to listen on
-    #[structopt(long, short, default_value = "3000")]
-    port: u16,
-    /// Backend to proxy in backend-name:host format (foo:foo.org)
-    #[structopt(long, short, parse(try_from_str = parse_key_value))]
-    backend: Vec<(String, String)>,
-    /// Edge dictionary in dictionary-name:key=value,key=value format
-    #[structopt(long, short, parse(try_from_str = parse_dictionary))]
-    dictionary: Vec<(String, HashMap<String, String>)>,
-    #[structopt(long)]
-    tls_cert: Option<PathBuf>,
-    #[structopt(long)]
-    tls_key: Option<PathBuf>,
-    /// Watch for changes to .wasm file, reloading application when relevant
-    #[structopt(long)]
-    watch: bool,
-}
 
 // re-writing uri to add host and authority. fastly requests validate these are present before sending them upstream
 fn rewrite_uri(
